move all the ExchangeRateTableData outside
- Noticed that the logic that fetches defaultCurrExchangeRates, invoke when on homepage, can be move inside the ExchangeRateTableData comp and run at once
- Might need to refactor ExchangeRateTable.js since it is really long
